requirements.txt
requests
mysql-connector-python
textblob
matplotlib
seaborn
python-dotenv
openai==0.28.0
flask